requirements.txt
django
django-filter
djangorestframework
django-cors-headers
djangorestframework-simplejwt
drf-yasg
gunicorn
python-dotenv
pyTelegramBotAPI
psycopg2-binary